AbstractHiearchyCalculator

+++ abstarct_calculator/Calculation/Archimetric.cs
@@ -4,6 +4,6 @@ namespace Calculation
 {
     public abstract class Archimetric
     {
-        public abstract decimal Calculation(int n, int m);
+        public abstract double Calculation(double n, double m);
     }
 }
