+++ abstarct_calculator/Sum_Multiply/Multiply.cs
@@ -7,9 +7,9 @@ namespace SumMultiply
 {
     public class Multiply : Archimetric
     {
-        public override decimal Calculation(int n, int m)
+        public override double Calculation(double n, double m)
         {
-            decimal result = n * m;
+            double result = n * m;
             return result;
         }
     }
